Fix notebook language info (#42)
This PR fixes the language info metadata in example notebooks. There might be a bug in Positron notebooks causing the incorrect language to be written to this field.

The incorrect languages are causing the wrong kernel to be started in a PR that reworks Positron notebook kernels: https://github.com/posit-dev/positron/actions/runs/12693575772, which this PR should fix.